⌛Refactor: 신고 개수 데이터 타입 수정
- Long -> Int로 수정

+++ src/main/kotlin/com/example/days/domain/user/model/User.kt
@@ -31,7 +31,7 @@ class User(
     @Column(name = "role")
     val role: UserRole,
 
-    @Column(name = "count_report") var countReport: Long = 0,
+    @Column(name = "count_report") var countReport: Int = 0,
 
 ): BaseEntity() {
